backend/requirements/base.txt
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
